requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
cffi==1.17.1
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.1.0
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
# easyocr==1.7.1
et-xmlfile==1.1.0
etelemetry==0.3.1
fastapi==0.115.0
fastjsonschema==2.20.0
fitz==0.0.1.dev2
fonttools==4.53.1
fqdn==1.5.1
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
# huggingface-hub==0.25.0
idna==3.10
imageio==2.35.1
ipywidgets==8.1.5
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
# jupyter==1.1.1
# jupyter-console==6.6.3
# jupyter-events==0.10.0
# jupyter-lsp==2.2.5
# jupyter_server==2.14.2
# jupyter_server_terminals==0.5.3
# jupyterlab==4.2.5
# jupyterlab_pygments==0.3.0
# jupyterlab_server==2.27.3
# jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-experimental==0.3.2
# langchain-huggingface==0.0.3
# langchain-ollama==0.1.0
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langgraph==0.2.35
langgraph-checkpoint==2.0.1
langsmith==0.1.134
lazy_loader==0.4
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mistune==3.0.2
# mkl-service==2.4.1
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
neo4j==5.24.0
nibabel==5.2.1
ninja==1.11.1.1
nipype==1.8.6
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
openai==1.45.1
openpyxl==3.1.5
orjson==3.10.7
overrides==7.7.0
pandas==2.2.2
pandocfilters==1.5.1
pathlib==1.0.1
# pdf2image==1.17.0
plotly==5.24.1
prometheus_client==0.20.0
protobuf==5.28.1
prov==2.0.1
pyarrow==17.0.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pydeck==0.9.1
pydot==3.0.1
pymongo==4.8.0
# PyMuPDF==1.24.10
# PyMuPDFb==1.24.10
pyparsing==3.1.4
pyproject-toml==0.0.10
# pytesseract==0.3.13
python-bidi==0.6.0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
pyvis==0.3.2
# pywin32
# pywinpty==2.0.13
pyxnat==1.6.2
rapidfuzz==3.9.7
rdflib==6.3.2
redis==5.0.8
referencing==0.35.1
regex==2024.9.11
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.20.0
# safetensors==0.4.5
# scikit-image==0.24.0
# scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==72.1.0
shapely==2.0.6
simplejson==3.19.3
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.38.5
streamlit==1.38.0
streamlit-option-menu==0.3.13
tenacity==8.5.0
terminado==0.18.1
thefuzz==0.22.1
threadpoolctl==3.5.0
tifffile==2024.8.30
# tiktoken==0.7.0
tinycss2==1.3.0
# tokenizers==0.19.1
toml==0.10.2
tqdm==4.66.5
traits==6.3.2
types-python-dateutil==2.9.0.20240906
typing-inspect==0.9.0
tzdata==2024.1
uri-template==1.3.0
uvicorn==0.30.6
watchdog==4.0.2
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
widgetsnbextension==4.0.13
yarl==1.11.1